feat(error): implement ErrorBoundary component with fallback rendering for improved error handling
- Added ErrorBoundary component to catch errors in child components and provide fallback UI.
- Integrated ErrorBoundary in HomePage, ArticlePage, and WritePage to enhance user experience during errors.
- Created specific fallback components for different sections, including DashboardRecentArticlesFallback, ArticleDisplaySectionFallback, and ArticleSidebarSectionFallback.
- Updated WritePage to utilize the new WriteEditorWorkspace component for better organization and error handling in the editor.
