+++ app/write/page.tsx
@@ -1,413 +1,17 @@
 'use client'
 
-import { useState, useEffect, useCallback, useRef } from 'react'
-import { useRouter, useSearchParams } from 'next/navigation'
-import { useEditor, EditorContent, JSONContent } from '@tiptap/react'
-import StarterKit from '@tiptap/starter-kit'
-import Underline from '@tiptap/extension-underline'
-import Link from '@tiptap/extension-link'
-import Placeholder from '@tiptap/extension-placeholder'
-import CodeBlockLowlight from '@tiptap/extension-code-block-lowlight'
-import TextAlign from '@tiptap/extension-text-align'
-import { lowlight } from '@/lib/lowlight'
-import { ResizableImage } from '@/components/editor/extensions/ResizableImage'
-import { EditorToolbar } from '@/components/editor/EditorToolbar'
-import { EditorActionBar } from '@/components/editor/EditorActionBar'
-import { ImageUploadDialog } from '@/components/editor/ImageUploadDialog'
-import { ExcerptTagsDialog } from '@/components/editor/ExcerptTagsDialog'
+import { useRouter } from 'next/navigation'
 import { useAuth } from '@/components/providers/AuthContext'
 import AppNavigation from '@/components/layout/AppNavigation'
 import { EditorChromeSkeleton } from '@/components/editor/EditorChromeSkeleton'
-import { useAutoSave } from '@/hooks/useAutoSave'
-import { useMutation } from 'convex/react'
-import { api } from '@/convex/_generated/api'
-import { useArticleById } from '@/hooks/convex'
-import type { Id } from '@/types/convex'
-import { toast } from 'sonner'
-import Image from 'next/image'
-import { EDITOR_PROSE_CLASS } from '@/lib/constants'
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from '@/components/ui/alert-dialog'
-
-const PUBLISH_EXCERPT_PREVIEW_MAX = 280
-
-const EMPTY_DOC: JSONContent = { type: 'doc', content: [] }
-
-type NavConfirmState = null | { kind: 'href'; href: string } | { kind: 'back' }
-
-function getInternalNavHref(
-  anchor: HTMLAnchorElement,
-  e: MouseEvent
-): string | null {
-  if (e.button !== 0) return null
-  if (e.metaKey || e.ctrlKey || e.shiftKey || e.altKey) return null
-  if (anchor.target === '_blank') return null
-  if (anchor.getAttribute('download') !== null) return null
-  const hrefAttr = anchor.getAttribute('href')
-  if (!hrefAttr || hrefAttr.startsWith('#')) return null
-  if (hrefAttr.startsWith('mailto:') || hrefAttr.startsWith('tel:')) return null
-  let url: URL
-  try {
-    url = new URL(hrefAttr, window.location.href)
-  } catch {
-    return null
-  }
-  if (url.origin !== window.location.origin) return null
-  if (
-    url.pathname === window.location.pathname &&
-    url.search === window.location.search
-  ) {
-    return null
-  }
-  return `${url.pathname}${url.search}${url.hash}`
-}
+import { ErrorBoundary } from '@/components/error/ErrorBoundary'
+import { EditorWorkspaceErrorFallback } from '@/components/error/SectionErrorFallback'
+import { WriteEditorWorkspace } from '@/components/editor/WriteEditorWorkspace'
 
 export default function WritePage() {
-  const [title, setTitle] = useState('')
-  const [excerpt, setExcerpt] = useState('')
-  const [tags, setTags] = useState('')
-  const [showExcerptTagsDialog, setShowExcerptTagsDialog] = useState(false)
-  const [coverImage, setCoverImage] = useState('')
-  const [showCoverImageDialog, setShowCoverImageDialog] = useState(false)
-  const [isPublishing, setIsPublishing] = useState(false)
-  const [articleId, setArticleId] = useState<string | undefined>()
-  const [editorContent, setEditorContent] = useState<JSONContent | null>(null)
-  const [hasUnsavedChanges, setHasUnsavedChanges] = useState(false)
-  const [navConfirm, setNavConfirm] = useState<NavConfirmState>(null)
-  const hasUnsavedRef = useRef(hasUnsavedChanges)
-  const [publishStatus, setPublishStatus] = useState<{
-    published: boolean
-    publishedAt: Date | null
-  }>({
-    published: false,
-    publishedAt: null,
-  })
-  const [publishConfirmOpen, setPublishConfirmOpen] = useState(false)
-
   const router = useRouter()
-  const searchParams = useSearchParams()
   const { isAuthenticated, isLoading } = useAuth()
 
-  useEffect(() => {
-    hasUnsavedRef.current = hasUnsavedChanges
-  }, [hasUnsavedChanges])
-
-  // Convex mutations
-  const createArticleMutation = useMutation(api.articles.createArticle)
-  const publishArticleMutation = useMutation(api.articles.publishArticle)
-  const deleteArticleMutation = useMutation(api.articles.deleteArticle)
-
-  // Initialize editor with proper configuration
-  const editor = useEditor({
-    immediatelyRender: false,
-    extensions: [
-      StarterKit.configure({
-        heading: {
-          levels: [1, 2, 3, 4, 5, 6],
-        },
-        codeBlock: false,
-      }),
-      Underline,
-      Link.configure({
-        openOnClick: false,
-        HTMLAttributes: {
-          class: 'underline cursor-pointer',
-        },
-      }),
-      ResizableImage.configure({
-        HTMLAttributes: {
-          class: 'max-w-full h-auto rounded-lg my-4',
-        },
-      }),
-      Placeholder.configure({
-        placeholder: 'Start writing your story...',
-      }),
-      CodeBlockLowlight.configure({
-        lowlight,
-        HTMLAttributes: {
-          class:
-            'rounded-lg bg-gray-900 text-gray-100 p-4 my-4 overflow-x-auto',
-        },
-      }),
-      TextAlign.configure({ types: ['heading', 'paragraph'] }),
-    ],
-    content: '',
-    editorProps: {
-      attributes: {
-        class: `${EDITOR_PROSE_CLASS} min-h-[400px] py-6 break-words`,
-      },
-    },
-    onCreate: ({ editor }) => {
-      // Set initial content state when editor is created
-      const json = editor.getJSON()
-      setEditorContent(json)
-    },
-    onUpdate: ({ editor }) => {
-      const json = editor.getJSON()
-      setEditorContent(json)
-      setHasUnsavedChanges(true)
-    },
-  })
-
-  // Auto-save hook - enable when we have a user and any content
-  const { isSaving, lastSavedAt, error, saveNow } = useAutoSave({
-    content: editorContent,
-    articleId,
-    title: title || 'Untitled',
-    excerpt,
-    tags: tags
-      .split(',')
-      .map((t) => t.trim())
-      .filter(Boolean),
-    coverImage: coverImage || undefined,
-    enabled: isAuthenticated && (hasUnsavedChanges || !!title),
-    onSaveSuccess: (response) => {
-      if (!articleId && response.id) {
-        setArticleId(response.id)
-      }
-      setHasUnsavedChanges(false)
-    },
-    onSaveError: (error) => {
-      console.error('Auto-save error:', error)
-    },
-  })
-
-  const draftIdParam = searchParams.get('id')
-
-  // Load draft using Convex query
-  const draft = useArticleById(
-    draftIdParam ? (draftIdParam as Id<'articles'>) : undefined
-  )
-
-  const savedArticleForLink = useArticleById(
-    articleId ? (articleId as Id<'articles'>) : undefined
-  )
-
-  // Load draft data when it arrives
-  useEffect(() => {
-    if (draft && editor) {
-      setArticleId(draft._id)
-      setTitle(draft.title)
-      setExcerpt(draft.excerpt || '')
-      setTags(draft.tags?.join(', ') ?? '')
-      setCoverImage(draft.coverImage || '')
-      setPublishStatus({
-        published: draft.published,
-        publishedAt: draft.publishedAt ? new Date(draft.publishedAt) : null,
-      })
-      if (draft.content) {
-        queueMicrotask(() => {
-          editor.commands.setContent(draft.content)
-        })
-        setEditorContent(draft.content)
-      }
-      // Reset unsaved changes flag after loading draft
-      setHasUnsavedChanges(false)
-    }
-  }, [draft, editor])
-
-  useEffect(() => {
-    const handler = (e: BeforeUnloadEvent) => {
-      const hasContent = !!editorContent
-      const hasMetadata = !!(title?.trim() || coverImage)
-      const autoSaveEnabled = isAuthenticated && (hasUnsavedChanges || !!title)
-      if (autoSaveEnabled && (hasContent || hasMetadata)) {
-        try {
-          const tagsArr = tags
-            .split(',')
-            .map((t) => t.trim())
-            .filter(Boolean)
-          localStorage.setItem(
-            'quilltip_draft_backup',
-            JSON.stringify({
-              title: title || 'Untitled',
-              content: editorContent ?? EMPTY_DOC,
-              excerpt,
-              tags: tagsArr.length ? tagsArr : undefined,
-              coverImage: coverImage || undefined,
-              articleId,
-              savedAt: Date.now(),
-            })
-          )
-        } catch {
-          // localStorage unavailable or full
-        }
-      }
-      if (hasUnsavedChanges) {
-        e.preventDefault()
-        e.returnValue = ''
-      }
-    }
-    window.addEventListener('beforeunload', handler)
-    return () => window.removeEventListener('beforeunload', handler)
-  }, [
-    hasUnsavedChanges,
-    isAuthenticated,
-    editorContent,
-    title,
-    coverImage,
-    excerpt,
-    tags,
-    articleId,
-  ])
-
-  useEffect(() => {
-    const onDocumentClick = (e: MouseEvent) => {
-      if (!hasUnsavedRef.current) return
-      const target = e.target
-      if (!(target instanceof Element)) return
-      const anchor = target.closest('a[href]')
-      if (!(anchor instanceof HTMLAnchorElement)) return
-      const href = getInternalNavHref(anchor, e)
-      if (!href) return
-      e.preventDefault()
-      e.stopPropagation()
-      setNavConfirm({ kind: 'href', href })
-    }
-    document.addEventListener('click', onDocumentClick, true)
-    return () => document.removeEventListener('click', onDocumentClick, true)
-  }, [])
-
-  useEffect(() => {
-    const handler = (e: KeyboardEvent) => {
-      if ((e.metaKey || e.ctrlKey) && e.key === 's') {
-        e.preventDefault()
-        saveNow()
-        setHasUnsavedChanges(false)
-      }
-    }
-    window.addEventListener('keydown', handler)
-    return () => window.removeEventListener('keydown', handler)
-  }, [saveNow])
-
-  const requestPublish = useCallback(() => {
-    if (!editor || editor.isEmpty) {
-      toast.warning('Please add content before publishing')
-      return
-    }
-    setPublishConfirmOpen(true)
-  }, [editor])
-
-  // Handle publish
-  const handlePublish = useCallback(async () => {
-    if (!editor || editor.isEmpty) {
-      setPublishConfirmOpen(false)
-      toast.warning('Please add content before publishing')
-      return
-    }
-    if (!editorContent) {
-      setPublishConfirmOpen(false)
-      toast.warning('Please add content before publishing')
-      return
-    }
-
-    setIsPublishing(true)
-    try {
-      // Save one final time before publishing
-      await saveNow()
-
-      let resultId: string
-
-      if (articleId) {
-        const published = await publishArticleMutation({
-          id: articleId as Id<'articles'>,
-        })
-        resultId = published.id
-      } else {
-        // Create and publish new article
-        resultId = await createArticleMutation({
-          title: title || 'Untitled',
-          content: editorContent,
-          excerpt: excerpt || undefined,
-          coverImage: coverImage || undefined,
-          tags: tags
-            .split(',')
-            .map((t) => t.trim())
-            .filter(Boolean),
-          published: true, // Publishing immediately
-        })
-      }
-
-      // If it was a new article, update the articleId
-      if (!articleId) {
-        setArticleId(resultId)
-      }
-
-      // Update publish status
-      setPublishStatus({
-        published: true,
-        publishedAt: new Date(),
-      })
-
-      toast.success('Article published successfully!')
-    } catch (error) {
-      console.error('Publish error:', error)
-      toast.error(
-        `Failed to publish: ${error instanceof Error ? error.message : 'Unknown error'}`
-      )
-    } finally {
-      setIsPublishing(false)
-    }
-  }, [
-    title,
-    editorContent,
-    excerpt,
-    coverImage,
-    tags,
-    saveNow,
-    articleId,
-    publishArticleMutation,
-    createArticleMutation,
-    editor,
-  ])
-
-  const handleDelete = useCallback(async () => {
-    if (!window.confirm('Are you sure you want to delete this draft?')) return
-    try {
-      if (articleId) {
-        await deleteArticleMutation({ id: articleId as Id<'articles'> })
-        router.push('/')
-      }
-    } catch (error) {
-      console.error('Delete error:', error)
-      toast.error(
-        error instanceof Error ? error.message : 'Failed to delete draft'
-      )
-    }
-  }, [articleId, deleteArticleMutation, router])
-
-  const handleBack = useCallback(() => {
-    if (hasUnsavedChanges) {
-      setNavConfirm({ kind: 'back' })
-    } else {
-      router.back()
-    }
-  }, [hasUnsavedChanges, router])
-
-  const confirmLeaveNavigation = useCallback(() => {
-    setNavConfirm((current) => {
-      if (!current) return null
-      const pending = current
-      queueMicrotask(() => {
-        if (pending.kind === 'href') {
-          router.push(pending.href)
-        } else {
-          router.back()
-        }
-      })
-      return null
-    })
-  }, [router])
-
-  // Authentication checks
   if (isLoading) {
     return (
       <div className="min-h-screen bg-background">
@@ -422,299 +26,16 @@ export default function WritePage() {
     return null
   }
 
-  const draftLoading = Boolean(draftIdParam) && draft === undefined
-  if (draftLoading) {
-    return (
-      <div className="min-h-screen bg-background">
-        <AppNavigation />
-        <EditorChromeSkeleton />
-      </div>
-    )
-  }
-
-  const tagsPreview =
-    tags
-      .split(',')
-      .map((t) => t.trim())
-      .filter(Boolean)
-      .join(', ') || 'No tags'
-  const excerptTrimmed = excerpt.trim()
-  const excerptPreview =
-    excerptTrimmed.length === 0
-      ? 'No excerpt'
-      : excerptTrimmed.length <= PUBLISH_EXCERPT_PREVIEW_MAX
-        ? excerptTrimmed
-        : `${excerptTrimmed.slice(0, PUBLISH_EXCERPT_PREVIEW_MAX).trimEnd()}...`
-
   return (
     <div className="min-h-screen bg-background">
       <AppNavigation />
-      <div className="flex flex-col pt-16">
-        {/* Action bar - full width, Back | Undo | Redo | Save | Preview | Publish */}
-        <EditorActionBar
-          editor={editor}
-          onBack={handleBack}
-          onSave={() => {
-            saveNow()
-            setHasUnsavedChanges(false)
-          }}
-          onPublish={requestPublish}
-          isSaving={isSaving}
-          error={error?.message ?? null}
-          isPublished={publishStatus.published}
-          isPublishing={isPublishing}
-          canPublish={!!editor}
-          lastSavedAt={lastSavedAt ?? undefined}
-          onDelete={handleDelete}
-          hasUnsavedChanges={hasUnsavedChanges}
-        />
-        <div className="flex-1 flex flex-col min-w-0 pb-8">
-          <div className="sticky top-16 z-40 bg-background w-full mb-6">
-            <EditorToolbar
-              editor={editor}
-              onFocusCoverImage={() => {
-                document
-                  .getElementById('field-cover-image')
-                  ?.scrollIntoView({ behavior: 'smooth', block: 'center' })
-                if (!coverImage) setShowCoverImageDialog(true)
-              }}
-              onFocusTitle={() => {
-                const el = document.getElementById(
-                  'article-title'
-                ) as HTMLInputElement | null
-                document
-                  .getElementById('field-article-title')
-                  ?.scrollIntoView({ behavior: 'smooth', block: 'center' })
-                el?.focus()
-              }}
-              onFocusExcerpt={() => setShowExcerptTagsDialog(true)}
-              onFocusTags={() => setShowExcerptTagsDialog(true)}
-            />
-            <div
-              className="absolute bottom-0 left-0 right-0 h-[2px] bg-sky-400 pointer-events-none"
-              aria-hidden
-            />
-          </div>
-          <div className="w-full max-w-4xl mx-auto px-4 sm:px-6">
-            {/* Cover Image */}
-            <div id="field-cover-image" className="mb-4">
-              {coverImage ? (
-                <div className="relative w-full h-56 sm:h-72 rounded-xl overflow-hidden group">
-                  <Image
-                    src={coverImage}
-                    alt="Article cover"
-                    fill
-                    sizes="(max-width: 768px) 100vw, 896px"
-                    className="object-cover"
-                  />
-                  <div className="absolute inset-0 bg-black/0 group-hover:bg-black/30 transition-colors flex items-center justify-center gap-3 opacity-0 group-hover:opacity-100">
-                    <button
-                      type="button"
-                      onClick={() => setShowCoverImageDialog(true)}
-                      className="px-4 py-2 bg-card text-foreground text-sm font-medium rounded-lg hover:bg-muted transition-colors shadow border border-border"
-                    >
-                      Change
-                    </button>
-                    <button
-                      type="button"
-                      onClick={() => {
-                        setCoverImage('')
-                        setHasUnsavedChanges(true)
-                      }}
-                      className="px-4 py-2 bg-card text-red-600 text-sm font-medium rounded-lg hover:bg-red-500/10 transition-colors shadow border border-border"
-                    >
-                      Remove
-                    </button>
-                  </div>
-                </div>
-              ) : (
-                <button
-                  type="button"
-                  onClick={() => setShowCoverImageDialog(true)}
-                  className="w-full h-28 border-2 border-dashed border-border rounded-xl flex items-center justify-center gap-2 text-muted-foreground hover:border-sky-400 hover:text-sky-500 transition-colors group"
-                >
-                  <svg
-                    className="w-5 h-5"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={1.5}
-                      d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z"
-                    />
-                  </svg>
-                  <span className="text-sm font-medium">Add cover image</span>
-                </button>
-              )}
-            </div>
-
-            <div id="field-article-title" className="mb-2">
-              <textarea
-                id="article-title"
-                value={title}
-                onChange={(e) => {
-                  setTitle(e.target.value)
-                  setHasUnsavedChanges(true)
-                }}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') {
-                    e.preventDefault()
-                    editor?.commands.focus()
-                  }
-                }}
-                placeholder="Untitled"
-                rows={1}
-                className="w-full resize-none overflow-hidden bg-transparent text-3xl font-semibold text-foreground placeholder:text-muted-foreground/50 focus:outline-none leading-snug py-2"
-              />
-              {savedArticleForLink?.authorUsername &&
-                savedArticleForLink.slug && (
-                  <p className="text-xs text-muted-foreground mt-1 font-mono">
-                    Public URL path (when published): /
-                    {savedArticleForLink.authorUsername}/
-                    {savedArticleForLink.slug}
-                  </p>
-                )}
-            </div>
-
-            {editor && (
-              <EditorContent
-                editor={editor}
-                className="editor-content min-h-[400px]"
-              />
-            )}
-
-            <div className="border-t border-border mt-8 pt-4">
-              <label
-                htmlFor="article-tags"
-                className="block text-xs font-medium text-muted-foreground mb-1"
-              >
-                Tags
-              </label>
-              <input
-                id="article-tags"
-                type="text"
-                value={tags}
-                onChange={(e) => {
-                  setTags(e.target.value)
-                  setHasUnsavedChanges(true)
-                }}
-                placeholder="Add tags separated by commas (e.g. rust, programming)"
-                className="w-full bg-transparent text-sm text-foreground placeholder:text-muted-foreground focus:outline-none py-1"
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <ImageUploadDialog
-        isOpen={showCoverImageDialog}
-        title="Add Cover Image"
-        onImageSelect={(url) => {
-          setCoverImage(url)
-          setHasUnsavedChanges(true)
-          setShowCoverImageDialog(false)
-        }}
-        onClose={() => setShowCoverImageDialog(false)}
-      />
-
-      <ExcerptTagsDialog
-        isOpen={showExcerptTagsDialog}
-        onClose={() => setShowExcerptTagsDialog(false)}
-        excerpt={excerpt}
-        tags={tags}
-        onExcerptChange={(v) => {
-          setExcerpt(v)
-          setHasUnsavedChanges(true)
-        }}
-        onTagsChange={(v) => {
-          setTags(v)
-          setHasUnsavedChanges(true)
-        }}
-      />
-
-      <AlertDialog
-        open={publishConfirmOpen}
-        onOpenChange={setPublishConfirmOpen}
+      <ErrorBoundary
+        fallback={({ reset }) => (
+          <EditorWorkspaceErrorFallback onRetry={reset} />
+        )}
       >
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>Publish this article?</AlertDialogTitle>
-            <AlertDialogDescription asChild>
-              <div className="space-y-3 text-left text-sm text-muted-foreground">
-                <p>
-                  Publishing stores your content on Arweave (permanent storage).
-                  You cannot undo this or remove that snapshot from Arweave.
-                </p>
-                <div className="space-y-2 rounded-md border border-border bg-muted/40 p-3 text-foreground">
-                  <div>
-                    <div className="text-xs font-medium uppercase tracking-wide text-muted-foreground">
-                      Title
-                    </div>
-                    <p className="mt-0.5 font-medium">
-                      {title.trim() || 'Untitled'}
-                    </p>
-                  </div>
-                  <div>
-                    <div className="text-xs font-medium uppercase tracking-wide text-muted-foreground">
-                      Tags
-                    </div>
-                    <p className="mt-0.5">{tagsPreview}</p>
-                  </div>
-                  <div>
-                    <div className="text-xs font-medium uppercase tracking-wide text-muted-foreground">
-                      Excerpt
-                    </div>
-                    <p className="mt-0.5 whitespace-pre-wrap break-words">
-                      {excerptPreview}
-                    </p>
-                  </div>
-                </div>
-              </div>
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel type="button">Cancel</AlertDialogCancel>
-            <AlertDialogAction
-              type="button"
-              onClick={() => {
-                setPublishConfirmOpen(false)
-                void handlePublish()
-              }}
-            >
-              Publish
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
-
-      <AlertDialog
-        open={navConfirm !== null}
-        onOpenChange={(open) => {
-          if (!open) setNavConfirm(null)
-        }}
-      >
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>Discard unsaved changes?</AlertDialogTitle>
-            <AlertDialogDescription>
-              Your draft is not saved to the server yet. If you leave now,
-              recent edits may be lost.
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel>Stay</AlertDialogCancel>
-            <AlertDialogAction
-              className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
-              onClick={confirmLeaveNavigation}
-            >
-              Leave
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
+        <WriteEditorWorkspace />
+      </ErrorBoundary>
     </div>
   )
 }
